🐛 fix(codex): address rich render review feedback
